ge: core: move filesystem to a separate lib
