+++ include/genesis/filesystem/file_content.h
@@ -1,7 +1,7 @@
 /*
  * BSD 3-Clause License
  *
- * Copyright (c) 2021-2022, Dmitry Shilnenkov
+ * Copyright (c) 2022, Dmitry Shilnenkov
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -32,17 +32,20 @@
 
 #pragma once
 
-#include <genesis/core/asserts.h>
-#include <genesis/core/bit.h>
-#include <genesis/core/defer.h>
-#include <genesis/core/enum.h>
-#include <genesis/core/exception.h>
-#include <genesis/core/export.h>
-#include <genesis/core/format.h>
-#include <genesis/core/interface.h>
-#include <genesis/core/log.h>
-#include <genesis/core/memory.h>
-#include <genesis/core/timestamp.h>
-#include <genesis/core/type_list.h>
-#include <genesis/core/utils.h>
-#include <genesis/core/version.h>
+#include <fstream>
+#include <iterator>
+
+namespace GE::FS {
+
+template<typename T>
+std::vector<T> readFile(const std::string& filepath)
+{
+    if (auto file = std::ifstream{filepath, std::ios::binary}; file) {
+        file >> std::noskipws;
+        return {std::istream_iterator<T>{file}, std::istream_iterator<T>{}};
+    }
+
+    return {};
+}
+
+} // namespace GE::FS
