+++ include/genesis/filesystem/tmp_dir_guard.h
@@ -1,7 +1,7 @@
 /*
  * BSD 3-Clause License
  *
- * Copyright (c) 2021-2022, Dmitry Shilnenkov
+ * Copyright (c) 2023, Dmitry Shilnenkov
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -32,17 +32,22 @@
 
 #pragma once
 
-#include <genesis/core/asserts.h>
-#include <genesis/core/bit.h>
-#include <genesis/core/defer.h>
-#include <genesis/core/enum.h>
-#include <genesis/core/exception.h>
 #include <genesis/core/export.h>
-#include <genesis/core/format.h>
-#include <genesis/core/interface.h>
-#include <genesis/core/log.h>
-#include <genesis/core/memory.h>
-#include <genesis/core/timestamp.h>
-#include <genesis/core/type_list.h>
-#include <genesis/core/utils.h>
-#include <genesis/core/version.h>
+
+#include <string>
+
+namespace GE::FS {
+
+class GE_API TmpDirGuard
+{
+public:
+    TmpDirGuard();
+    ~TmpDirGuard();
+
+    std::string_view path() const { return m_path; }
+
+private:
+    std::string m_path;
+};
+
+} // namespace GE::FS
