+++ src/genesis/core/CMakeLists.txt
@@ -1,7 +1,6 @@
 set(INCLUDE_DIR ${GE_INCLUDE_DIR}/genesis/core)
 
 list(APPEND CORE_SOURCES
-    filesystem.cpp
     log.cpp
     )
 
@@ -12,7 +11,6 @@ list(APPEND CORE_HEADERS
     ${INCLUDE_DIR}/enum.h
     ${INCLUDE_DIR}/exception.h
     ${INCLUDE_DIR}/export.h
-    ${INCLUDE_DIR}/filesystem.h
     ${INCLUDE_DIR}/format.h
     ${INCLUDE_DIR}/hash.h
     ${INCLUDE_DIR}/interface.h
@@ -27,7 +25,6 @@ ge_add_module(core
     SOURCES ${CORE_SOURCES} ${CORE_HEADERS}
     INCLUDE_DIRS ${INCLUDE_DIR}
     PUBLIC_DEPS
-        Boost::filesystem
         Boost::mpl
         fmt::fmt
         magic_enum::magic_enum
